requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
langchain-chroma
openai
chromadb
pydantic
asyncpg
pytest
pytest-asyncio
httpx
python-dotenv
docker
python-multipart
unstructured[all-docs]


